docs/local-dev/gemini_bundle.py
import google.generativeai as genai
import pandas as pd
import os
from typing import Dict, List, Any

# Configure Gemini API
genai.configure(api_key=os.getenv('GEMINI_API_KEY'))

def parse_excel_with_gemini(file_path: str) -> Dict[str, Any]:
    """
    Parse complex Excel file using Gemini AI for intelligent data extraction.
    Handles V.24 format with multiple sheets and specimen data.
    """
    try:
        xl = pd.ExcelFile(file_path)
        model = genai.GenerativeModel('gemini-pro')

        parsed_data = {
            'file_info': {
                'filename': os.path.basename(file_path),
                'sheets': len(xl.sheet_names),
                'sheet_names': xl.sheet_names
            },
            'specimens': [],
            'metadata': {}
        }

        # Process each sheet
        for sheet_name in xl.sheet_names:
            df = xl.parse(sheet_name)

            # Use Gemini to analyze and extract structured data
            prompt = f"""
            Analyze this Excel sheet data and extract specimen information.
            Sheet name: {sheet_name}
            Columns: {list(df.columns)}
            First 10 rows: {df.head(10).to_string()}

            Extract:
            - Specimen details (ID, taxon, collector, etc.)
            - PCR reactions
            - Laboratory information
            - Collection data

            Return as structured JSON.
            """

            response = model.generate_content(prompt)
            # Parse response and add to parsed_data
            # (Implementation would depend on Gemini response format)

        return parsed_data

    except Exception as e:
        print(f"Error parsing Excel: {e}")
        return {}

if __name__ == "__main__":
    # Example usage
    result = parse_excel_with_gemini("data/data.xlsx")
    print(result)