docs/local-dev/requirements.txt
# Requirements for local Python scripts
pandas
openpyxl
requests
google-generativeai